Renamed names and cleaned up code
