+++ Masquerade4997Code/src/main/java/org/firstinspires/ftc/teamcode/Robots/Robot1/Prototype/Autonomous.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Robots.Robot1.Bobot;
+package org.firstinspires.ftc.teamcode.Robots.Robot1.Prototype;
 
 import Library4997.MasqWrappers.MasqLinearOpMode;
 
@@ -6,7 +6,7 @@ import Library4997.MasqWrappers.MasqLinearOpMode;
  * Created by Keval Kataria on 9/15/2019
  */
 public class Autonomous extends MasqLinearOpMode {
-    BobotRobot robot = new BobotRobot();
+    PrototypeRobot robot = new PrototypeRobot();
     @Override
     public void runLinearOpMode() throws InterruptedException {
         while(!opModeIsActive()) {
