+++ Masquerade4997Code/src/main/java/org/firstinspires/ftc/teamcode/Robots/Robot1/Prototype/Teleop.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Robots.Robot1.Bobot;
+package org.firstinspires.ftc.teamcode.Robots.Robot1.Prototype;
 
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
@@ -13,7 +13,7 @@ public class Teleop extends MasqLinearOpMode {
     private double twisterPosition = 0;
     @Override
     public void runLinearOpMode() throws InterruptedException {
-        BobotRobot robot = new BobotRobot();
+        PrototypeRobot robot = new PrototypeRobot();
 
         while(!opModeIsActive()) {
             dash.create("Done lol");
@@ -24,10 +24,10 @@ public class Teleop extends MasqLinearOpMode {
         while(opModeIsActive()) {
             robot.MECH(controller1);
             robot.Grabber.setPosition(grabberPosition);
-            if (controller1.a()) grabberPosition+= 0.1;
+            if (controller1.aOnPress()) grabberPosition+= 0.1;
             robot.Twister.setPosition(twisterPosition);
-            if (controller1.b()) twisterPosition+= 0.1;
-            robot.Gripper.setPower(controller1.leftTrigger()-controller1.rightTrigger());
+            if (controller1.bOnPress()) twisterPosition+= 0.1;
+            robot.Intake.setPower(controller1.leftTrigger()-controller1.rightTrigger());
         }
     }
 }
